requirements.txt
certifi==2022.6.15
charset-normalizer==2.0.12
contourpy==1.1.1
cycler==0.12.0
et-xmlfile==1.1.0
fonttools==4.43.0
future==1.0.0
gitdb==4.0.11
GitPython==3.1.43
idna==3.3
importlib-resources==6.1.0
kiwisolver==1.4.5
matplotlib==3.8.0
numexpr==2.10.1
numpy==1.23.0
openpyxl==3.0.10
packaging==23.2
pandas==2.2.2
pandastable==0.13.1
Pillow==10.0.1
pyasn1==0.6.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.0
rsa==4.9
six==1.16.0
smmap==5.0.1
tzdata==2024.1
urllib3==1.26.9
xlrd==2.0.1
XlsxWriter==3.1.9
zipp==3.17.0
